feat: 아티클 조회 페이지를 기본 화면으로 설정
